requirements.txt
asn1crypto==0.24.0
Babel==2.7.0
bcrypt==3.1.7
blinker==1.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
crcmod==1.7
cryptography==2.7
Flask==1.0.3
Flask-Babel==0.12.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-OAuthlib==0.9.5
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-WTF==0.14.2
geoip2==2.9.0
gevent==1.4.0
greenlet==0.4.15
has==0.0.4
idna==2.7
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
lxml==4.3.4
Markdown==2.6.11
MarkupSafe==1.1.1
maxminddb==1.4.1
oauthlib==2.1.0
olefile==0.46
paramiko==2.6.0
Pillow==4.3.0
psutil==5.6.3
pyasn1==0.4.5
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.8.2
PyJWT==1.7.1
pymongo==3.8.0
PyNaCl==1.3.0
pytz==2019.1
PyYAML==5.1.1
redis==2.10.6
regex==2019.6.8
requests==2.20.1
requests-oauthlib==1.1.0
six==1.12.0
tld==0.9.3
urllib3==1.24.3
Werkzeug==0.15.4
WTForms==2.2.1
wordcloud==1.5.0
nltk==3.4.4